refactor: image path change
