+++ src/components/Modal.jsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components'
-import logo from '../assets/logo.svg'
+import logo from '../assets/icons/logo.svg'
 import { useNavigate } from 'react-router-dom'
 
 const Modal = () => {
